First upload - Basic framework
The basic framework of the website and the backend. It is able to grab data from the RIOT API and display it on the website. Requires a .env file with a riot api key, port, and GET url. Also, these files need to be in an express.js project with a few extra add-ons like cors and dotenv installed.